Added a new api `NewXrayRequest`
Added a new api `NewXrayRequest` which will allow us to create a base64 encoded request from given dat file dir and config path in the function argument. I think it's a better a more formal way to create a Xray Run Request which is straight forward. But before it was little annoying to me as I need to first create a Xray request pass paths my self then encode it myslelf then pass to Run funciton etc blah blah alot of work to me. So this function does all of that with a single call which I feel is useful. Also it reutns a error if sometime broke in between. If you have any other questions about the implementation reply to this thread