Changed it to min api 21
Changed it to min api 21 cause you see there are alot of devices which runs xray and are less than api level 28. So if you set it to 28 then alot of devices gets cutt of  support like few days ago I faced the same issue when I tried to build libXray and setup in android project it just don't work on devices below api level 28 so made it 21 so it works on android 5.0 and up devices. For reference v2rayNG also uses same api level for xray wrapper which is 21