Xray : Add new api to get xray state.
So in this pr I have added a new api for getting the core state running or not. I looked into the `Instance` struct in the `Xray-Core` core/xray.go file and found there is a running bool which returns if its running or not. So the issue was it was not exporting runnning. So first I did a contribution in main Xray core to add that feature which is now merged [#447](https://github.com/XTLS/Xray-core/pull/4775) and then I added the same thing here in the client wrapper. and it looks like this

```go
// Get Xray State
func GetXrayState() bool {
	return coreServer.IsRunning()
}
```