Build : Add api level set option
So i felt before that its a useful feature cause someone might want to decide their own api level not everyone wants to use 21 I see. so added a cmd line argument in the build script to make it work. If not provided by default it will use 21. Also udpated the readme